feat(loadtest): standalone loadtest binary with whip/whep/channel
New `loadtest` binary with three subcommands:

- whip: N concurrent synthetic WHIP publishers (requires rsmpeg)
- whep: N concurrent WHEP subscribers against one stream
- channel: DataChannel <-> UDP throughput/latency/bidirectional bench,
  self-contained with an in-process liveion (requires source)

The datachannel_loadtest logic moves to src/bin/loadtest_channel so both
binaries share it; datachannel_loadtest stays as a thin wrapper that
keeps its positional-mode + LOADTEST_* env-var interface. Subcommands
whose feature is disabled fail with a clear rebuild hint.
